objects and so on ...

+++ semestralka/lib/giraffe/instance.rb
@@ -8,6 +8,8 @@ module Giraffe
 
         def initialize(clz)
             @clz = clz
+            @s_instance = nil   # instance super_class
+            @env = nil
         end        
 
         # Vytvori novou vrstvu Env - pokud je 
@@ -17,42 +19,58 @@ module Giraffe
         # metod a obecne veskery kod, ktery se 
         # nachazi v tele tridy ... 
         def constructor
-               
+
             super_class = @clz[2]
 
             # pokud mam super_class
             if super_class != nil 
 
                 # vyzvedni deklaraci super_class
-                s_clz = @clz[3].cls(super_class)
+                return_value, msg = @clz[3].cls(super_class)
+                case msg  
+                when nil;
+                when :error then return return_value + "\n\tin #{@clz[0]} constructor", msg
+                else return return_value, msg
+                end
+
+                # ziskani tridy probehlo uspesne
+                s_clz = return_value
 
                 dbg("constructing superclass '#{s_clz[0]}'",:Instance)
 
                 # vytvor instanci super_class
                 # ale uz ji nepredavej argumenty
-                s_instance = Instance.new(s_clz)
-                return_value, msg = s_instance.constructor
-                return return_value, msg if msg != nil
+                @s_instance = Instance.new(s_clz)
+                return_value, msg = @s_instance.constructor
+                case msg  
+                when nil;
+                when :error then return return_value + "\n\tin #{@clz[0]} constructor", msg
+                else return return_value, msg
+                end
 
                 # postav novou vrstvu Env na vrstve
                 # na ktere je postaven super_class
-                @env = Env.new(s_instance.env)
+                @env = Env.new(@s_instance.env)
 
                 # TODO ... mam davat s_instance do pool ?
                 # asi bych mel :)
-    
+
             else
 
                 @env = Env.new
 
             end
-        
+
             # vykonej instrukce (zavedeni metod apod.)
             for i in @clz[1] do
                 return_value, msg = i.run(@env)
-                return return_value, msg if msg != nil
+                case msg  
+                when nil;
+                when :error then return return_value + "\n\tin #{@clz[0]} constructor", msg
+                else return return_value, msg
+                end
             end
-            
+
             return nil, nil
 
         end
@@ -60,24 +78,35 @@ module Giraffe
         def method_call(id, args, call_env)
 
             dbg("method_call for method '#{id}'",:Instance)
-            
+
             # vyzvedni deklaraci 
-            func = @env.func(id,args)
+            func, msg = @env.func(id,args)
+            case msg  
+            when nil;
+                # pokud se hledal iniciator (metoda se stejnym nazvem jako je 
+                # nazev tridy, ignoruj, ze neexistuje 
+            when :error then return id == @clz[0] ? [nil, nil] : [func, msg]
+            else return func, msg
+            end
 
             # vytvor novou vrstvu prostredi
             # ale na vrstve kde byla deklarovana
-            # funkce, nikoliv na vrstve volani
-            # TODO env zdroj ... odkud ? Inheritance x vnorene funkce ...
-            # chtelo by to neco jako kontrolu "level of Env" a vybrat max
-            new_env = Env.new(@env)     # dedene funkce uvidi do obsahu tridy potomka
-                                        # @env se tak ale stava nejvyssi vrstvou, takze 
-                                        # nebudou fungovat vnorene funkce
+            # metody, nikoliv na vrstve volani
+            # 
+            # Tady je potreba zjistit co je vyssi - jestli vrstva metody,
+            # ze ktere se vola nebo vrstva vrstva potomka
+            #
+            # Dedene metody totiz musi byt povysene na Env potomka
+            # ale lokalni metody zase by touhle upravou klesli pokud
+            # jsou vnorene v jinych metodach
+            #
+            # Takze - pokud volam z potomka, urcite beru jako vychozi env
+            # definicni env metody
+            # 
+            # Pokud volam z predka, beru jako vychozi env potomka
+            new_env = Env.new(@env.level < func[2].level ? func[2] : @env)  
 
-            #new_evn = Env.new(func[2]) # funkce ve funkcich v potomkovi uvidi vsude po 
-                                        # tride jak maji, ale dedene funkce nevidi do 
-                                        # potomka 
-
-            return_value, msg = inner_run(new_env,call_env,func,args)
+            return_value, msg = inner_run(new_env,call_env,func,args,id)
 
             # zavri vrstvu
             new_env.destroy
@@ -88,9 +117,9 @@ module Giraffe
 
         def destructor
         end
-        
+
         private
-        def inner_run(env,call_env,func,args)
+        def inner_run(env,call_env,func,args,id)
 
             # deklarace
             return_value = msg = nil
@@ -115,7 +144,10 @@ module Giraffe
                 case msg
                 when :return then return return_value, nil
                 when :exit then return return_value, msg
-                when :break then raise "Unexpected break"
+                when :error then return return_value + "\n\tin method '#{id}'", msg
+                when :break then return orange("Unexpected break") + "\n\tin method '#{id}'", :error
+                when nil;
+                else return return_value, msg
                 end
             end
